requirements.txt
alembic==1.7.7
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.32
urllib3==1.26.9
zipp==3.7.0
